# frontend/apidoc/openapi.yaml
openapi: 3.0.3
info:
  title: 'Codechallenge Cart API'
  description: 'Codechallenge Cart API'
  version: 1.0.0
servers:
  -
    url: 'http://localhost:8005'
paths:
...
  /api/user:
    get:
      summary: 'GET api/user'
      operationId: gETApiuser
      description: 'Return a Json array containing the information of the user'
      parameters:
        -
          in: header
          name: Authorization
          description: ''
          example: 'Bearer N7fp6GTjO9CJD1QIhqv0Ty1ZZbJeS3tFIbToFJZQ'
          schema:
            type: string
      responses:
        200:
          description: success
          content:
            application/json:
              schema:
                type: object
                example:
                  name: Javier
                  email: javier@smartidea.es
                  address: 'Mi dirección'
                  registered: true
                properties:
                  name:
                    type: string
                    example: Javier
                  email:
                    type: string
                    example: javier@smartidea.es
                  address:
                    type: string
                    example: 'Mi dirección'
                  registered:
                    type: boolean
                    example: true
        401:
          description: unauthenticated
          content:
            application/json:
              schema:
                type: object
                example:
                  message: Unauthenticated
                properties:
                  message:
                    type: string
                    example: Unauthenticated
      tags:
        - 'User management'
  /api/user/login:
    post:
      summary: 'POST api/user/login'
      operationId: pOSTApiuserlogin
      description: "Return a json array containing a new auth token for registered user.\nChecks if the email belongs to a existing domain"
      parameters:
        -
          in: header
          name: Authorization
          description: ''
          example: 'Bearer N7fp6GTjO9CJD1QIhqv0Ty1ZZbJeS3tFIbToFJZQ'
          schema:
            type: string
      responses:
        200:
          description: success
          content:
            application/json:
              schema:
                type: object
                example:
                  token: 34|0JJwyg6oCkEBnlAGFVOlq5f42SY9u476JSCBVUwT
                properties:
                  token:
                    type: string
                    example: 34|0JJwyg6oCkEBnlAGFVOlq5f42SY9u476JSCBVUwT
        401:
          description: unauthenticated
          content:
            application/json:
              schema:
                type: object
                example:
                  message: Unauthenticated
                properties:
                  message:
                    type: string
                    example: Unauthenticated
        403:
          description: 'wrong credentials'
          content:
            application/json:
              schema:
                type: object
                example:
                  message: 'Login error'
                properties:
                  message:
                    type: string
                    example: 'Login error'
        422:
          description: 'invalid parameters'
          content:
            application/json:
              schema:
                type: object
                example:
                  message: 'The password field is required.'
                properties:
                  message:
                    type: string
                    example: 'The password field is required.'
      tags:
        - 'User management'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  description: 'valid email from existing domain.'
                  example: codechallenge@gmail.com
                password:
                  type: string
                  description: ''
                  example: et
              required:
                - email
                - password
  /api/user/register:
    post:
      summary: 'POST api/user/register'
      operationId: pOSTApiuserregister
      description: "Process a registration request.\nChecks if the email belongs to a existing domain"
      parameters:
        -
          in: header
          name: Authorization
          description: ''
          example: 'Bearer N7fp6GTjO9CJD1QIhqv0Ty1ZZbJeS3tFIbToFJZQ'
          schema:
            type: string
      responses:
        200:
          description: success
          content:
            text/plain:
              schema:
                type: string
                example: ''
        401:
          description: unauthenticated
          content:
            application/json:
              schema:
                type: object
                example:
                  message: Unauthenticated
                properties:
                  message:
                    type: string
                    example: Unauthenticated
        403:
          description: 'register failed'
          content:
            application/json:
              schema:
                type: object
                example:
                  message: 'Register error'
                properties:
                  message:
                    type: string
                    example: 'Register error'
        422:
          description: 'invalid parameters'
          content:
            application/json:
              schema:
                type: object
                example:
                  message: 'The password field is required.'
                properties:
                  message:
                    type: string
                    example: 'The password field is required.'
      tags:
        - 'User management'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  description: 'valid email from existing domain.'
                  example: codechallenge@gmail.com
                password:
                  type: string
                  description: ''
                  example: laudantium
                name:
                  type: string
                  description: ''
                  example: iure
                address:
                  type: string
                  description: ''
                  example: maiores
              required:
                - email
                - password
                - name
                - address
  /api/user/orders:
    get:
      summary: 'GET api/user/orders'
      operationId: gETApiuserorders
      description: 'Return a Json array containing the orders of the user'
      parameters:
        -
          in: header
          name: Authorization
          description: ''
          example: 'Bearer N7fp6GTjO9CJD1QIhqv0Ty1ZZbJeS3tFIbToFJZQ'
          schema:
            type: string
      responses:
        200:
          description: success
          content:
            application/json:
              schema:
                type: object
                example:
                  orders:
                    -
                      id: 6
                      user_id: 104
                      address: 'Mi dirección'
                      total: 35
                      created_at: '2023-04-10T03:31:10.000000Z'
                      updated_at: '2023-04-10T03:31:10.000000Z'
                properties:
                  orders:
                    type: array
                    example:
                      -
                        id: 6
                        user_id: 104
                        address: 'Mi dirección'
                        total: 35
                        created_at: '2023-04-10T03:31:10.000000Z'
                        updated_at: '2023-04-10T03:31:10.000000Z'
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                          example: 6
                        user_id:
                          type: integer
                          example: 104
                        address:
                          type: string
                          example: 'Mi dirección'
                        total:
                          type: integer
                          example: 35
                        created_at:
                          type: string
                          example: '2023-04-10T03:31:10.000000Z'
                        updated_at:
                          type: string
                          example: '2023-04-10T03:31:10.000000Z'
        401:
          description: unauthenticated
          content:
            application/json:
              schema:
                type: object
                example:
                  message: Unauthenticated
                properties:
                  message:
                    type: string
                    example: Unauthenticated
      tags:
        - 'User management'
...
components:
  securitySchemes:
    default:
      type: http
      scheme: bearer
      description: ''
security:
  -
    default: []